Move lower_expr_if to a separate file.
commit-id:014f03cc
